Add new utility types for new system:
Using a tuple instead of a union of possible events:

type MyEvents = MyEvent1 | MyEvent2

becomes

type MyEvents = [MyEvent1, MyEvent2]

This is because it's hard / not a good idea to convert a union into a tuple.

It is, however, easy to convert a tuple into a union (not strictly needed though).